requirements.txt
networkx==3.4.2
numpy==2.3.3
pandas==2.3.3
pymetis==2025.2.1
PyYAML==6.0.3
torch==2.7.0+cu128
tqdm==4.67.1
